pyproject.toml
[tool.poetry]
name = "Avatar-Embed"
version = "0.1.0"
description = "为头像嵌入图标的小工具"
authors = ["AzideCupric <rukuy@qq.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Enraged-Dun-Cookie-Development-Team/Avatar-Embed"
repository = "https://github.com/Enraged-Dun-Cookie-Development-Team/Avatar-Embed"
documentation = "https://github.com/Enraged-Dun-Cookie-Development-Team/Avatar-Embed/README.md"
keywords = []
classifiers = []
packages = [{include = "avatar_embed"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.0.0"
httpx = "^0.24.1"
click = "^8.1.7"

[tool.poetry.group.lint.dependencies]
ruff = "^0.0.275"
isort = "^5.10.1"
black = "^23.1.0"
nonemoji = "^0.1.2"
pre-commit = "^3.1.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
pytest-mock = "^3.11.1"
respx = "^0.20.2"

[tool.black]
preview = true
line-length = 88
target-version = ["py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 88
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["C901"]

line-length = 88
target-version = "py310"

[tool.ruff.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
